+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-alpha11] - 2019-03-14
+### Fix
+ - Fixed `modify` function in `UploadService` class.
+ - Removed functionality of `copy` function from `UploadService` class.
+
 ## [1.0.0-alpha10] - 2019-03-12
 ### Fix
  - `getToArrayAttributes()` function added to `AccessToken` and `Target` classes.